:space_invader: finished base GA refactor
